pyproject.toml
[project]
name = "cr2ad-experiments"
version = "0.1.0"
description = "CR2AD decoding experiments with Hydra"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.1.0",
  "transformers>=4.38.0",
  "datasets>=2.16.0",
  "hydra-core>=1.3.2",
  "wandb>=0.16.0",
  "optuna>=3.5.0",
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "matplotlib>=3.8.0",
  "seaborn>=0.13.0",
  "scipy>=1.10.0",
  "scikit-learn>=1.3.0",
  "tqdm>=4.66.0"
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
